Allow auto-refresh to work with searches
A new piece of state is added, the 'mode'. The mode is used to determine
if we are in search mode or normal mode.

  * The 'normal' mode is selected at start-up
  * The mode changes to 'search' when the enter key is pressed in a
    non-empty search bar.
  * The mode returns to normal when the search bar is emptied.

The refresh_data function will use the mode to know if it should collect
search-result data, or normal data.

Note: The enter key is not needed to change the search term if already
in search mode (eg, changing the text by highlighting instead of
clearing the field first).